backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
audioop-lts==0.2.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
elevenlabs==2.40.0
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.30.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
reportlab==4.4.10
requests==2.33.0
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
